✨ feat(task): chore shortcut
Allows to enter more quickly any chore if it doesn't exist,
or add +chore tag and remove -todo tag if it exists.

helper script task-autotag abstracts away all the logic,
so that chore.bash has only to specify the +chore tag

chores should have a higher priority, so an urgency
coefficient is set up for the user tag +chore.